compositor: Optimize repaint scheduling for fullscreen surfaces.
This is an attempt at a "unredirect fullscreen window"
style optimization to reduce latency / allow for relaxed
repaint window deadlines if a direct scanout capable
fullscreen surface is displayed in the primary scanout.

This patch introduces to repaint-window deadlines,
a conservative early deadline if composition/rendering
is needed, and a late deadline if rendering can be
skipped because a fullscreen surface takes over the
primary scanout. The late deadline allows a fullscreen
client to render and commit its updates closer to the
start of the next target frame, thereby allowing for
more complex rendering to be done without skipping frames,
and for lower latency in presentation.

It works as follows:

First weston_output_finish_frame() checks if a repaint
of the output is needed and exits the repaint loop early
if this isn't the case. This allows an instant restart
of the repaint loop on most backends once a first surface
update gets committed for the output, thereby calling
weston_output_finish_frame() again.

If a repaint is needed, plane assignments are performed
early in weston_output_prepare_repaint(). If a surface
view is assigned to the primary scanout plane then the
composition surface will be occluded/not displayed for
the next output repaint and therefore slow rendering
into that surface can be skipped, allowing for a faster
repaint of the output.

If rendering is needed, we pick the early repaint-window
deadline as in the past from the 'repaint-window' parameter,
leaving enough headroom for rendering during the repaint.

If no rendering is needed then a fast repaint allows to
select a repaint deadline closer to the vblank of the
next frame, chosen from the new compositor parameter
'late-repaint-window'.

After picking the early or late deadline we check if that
deadline can be still met. We reschedule the repaint for
the next frames early deadline if the selected deadline
can't be met due to a too late restart of the repaint loop.

If the deadline can be met, set the repaint timer accordingly
for early or late repaint. In the timer callback we must recheck
the plane assignment as it might have changed during the repaint
delay due to new surfaces committing damage.

This allows to collect/process surface updates as close as
possible to the target frames vblank. The optimization only
works if a fullscreen surface takes over the primary scanout
though, thereby avoiding the need for any rendering  to handle
damage on the outputs composition surface.

Single display testing on a 5 years old Intel HD Ironlake mobile
suggests this optimization works reasonably well.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>