compositor-drm: Use 0 vblank wait for overlay plane updates.
This avoids a frame of extra lag due to vblank events being
queued for 1 vblank into the future after the vblank when
the overlay planes have updated. This is needed for kms
drivers where a plane update always incurs blocking until
the target vblank of the plane update has happened.

This gives correct behaviour for intel-kms on at least
Linux 3.19-4.2. However, different drivers will behave
inconsistently wrt. blocking one vblank or not, so this
is just a hack to make overlays on intel work timing-wise
for realistic testing of a following patch which optimizes
hardware overlay presentation.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>