compositor-drm: Use fast repaint with hardware overlays if possible.
If hardware overlays are used for non-fullscreen direct
scanout capable surfaces, try to allow for a fast repaint
with a late composition deadline. Check if the hardware
overlays cover all damaged regions of the outputs
composition surface. If this is the case then the damage
won't be visible onscreen, so rendering for damage handling
can be skipped for a fast low-latency repaint at the late
repaint deadline.

This would work for a mostly static desktop, except for
some hardware overlay which may display a video or game
or other dynamic content. It will cause a frame skip
if some visible unoccluded damage arrives for the composition
surface between the decision for fast vs. slow repaint
and the actual repaint delay timer firing, so this is
a bit of a niche optimization.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>